[Refactor] Re-build file structure
1. add ReadMe.md
2. Re-build
