+++ README.md
@@ -1,0 +1,19 @@
+##English
+###Global Linear Model for NLP Part-of-Speech
+####Run Demo
+```bash
+$ sh demo.sh
+```
+####References:
+* [NLP-GLM-Tagger](https://github.com/JimJarvis/NLP-GLM-Tagger) Data Strucutre Definition and Feature Template
+* [NLP Courses at Columbia University](http://www1.cs.columbia.edu/nlp/index.cgi)
+
+##中文
+###全局限性模型解决NLP中的词性标注问题
+####运行 Demo
+```bash
+$ sh demo.sh
+```
+####参考：
+* [NLP-GLM-Tagger](https://github.com/JimJarvis/NLP-GLM-Tagger) 中的数据格式的定义
+* [哥伦比亚大学NLP课程的视频和课件](http://www1.cs.columbia.edu/nlp/index.cgi)
