+++ src/demo.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+python train.py
+python predict.py
